Create README.md
Sentiment Analysis is becoming one of the most profound research areas for prediction and classification. Automated sentiment analysis of text is used in fields where products and services are reviewed by customers and critics. Thus, sentiment analysis becomes important for businesses to draw a general opinion about their products and services. In this particular project, I am doing Sentiment Analysis of the messy data of IMDb which hides a lot of information under the hood waiting to be discovered. This analysis will help find opinions of people about movies from their reviews, if it is positive or negative. One can in turn formulate a public opinion about a movie.